Support decimal amounts

+++ app/src/main/java/xyz/travitia/lister/data/model/Item.kt
@@ -3,7 +3,7 @@ package xyz.travitia.lister.data.model
 data class Item(
     val id: Int,
     val name: String,
-    val amount: Int?,
+    val amount: Double?,
     val amountUnit: String?,
     val inCart: Boolean,
     val list: Int,
@@ -12,14 +12,14 @@ data class Item(
 
 data class CreateItemRequest(
     val name: String,
-    val amount: Int? = null,
+    val amount: Double? = null,
     val amountUnit: String? = null,
     val category: String? = null
 )
 
 data class UpdateItemRequest(
     val name: String,
-    val amount: Int?,
+    val amount: Double?,
     val amountUnit: String?,
     val category: String?
 )
